fix: update firebase sdk installation
